Start over using direct dump from event stream as input
Record the macro with, for example:

$ cat /dev/input/event0 > moves

And playback with:

$ python grinder.py moves

The event stream contains time data, and is used to delineate
commands. So, grinder.py can read this data, convert the timestamp
data into seconds to sleep, and write out the original data at
the original intervals back to /dev/input/event0.